requirements.txt
     discord.py
     atproto
     python-dotenv
     requests